Cleanup: Removed confusing JMetrics.h
